graphql(v2): share code between orders root query and account.orders
